% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task13.R
\name{Task13}
\alias{Task13}
\title{Task13}
\usage{
bootstrap(n,X,Y)
}
\arguments{
\item{n}{number of tests}

\item{x}{is the a random variable}

\item{y}{is the second random varaible}
}
\value{
Return a vector with 6 outputs. The true correlation, the estimated correlation, the standard error of the estimated correlation, the bias, and a 95% confidence interval.
}
\description{
This is a function to process two random variables in relation to their correlation.
}
\author{
Nis Klausen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{nklau1414@student.sdu.dk} \cr
}
\keyword{boot}
\keyword{bootstrap}
\keyword{coorelation}

