% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task2.R
\name{Task2}
\alias{Task2}
\title{PackageExam2016}
\usage{
Task2(n, l = 1, d = 1)
}
\arguments{
\item{n}{number of tests}

\item{l}{is the length of the needle}

\item{d}{is the distance between the lines}
}
\value{
Yet to be decided
}
\description{
This is a function to estimate pi using buffon's needle
}
\author{
Nis Klausen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{nklau1414@student.sdu.dk} \cr
}
\keyword{Buffon}
\keyword{Needle}
\keyword{Pi}

