% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task31.R
\name{Task31}
\alias{Task31}
\title{Task31}
\usage{
Task31(x, df, n = 1000)
}
\arguments{
\item{x}{the variable that the q-square test is done upon.}

\item{df}{degrees of freedom}

\item{n}{number of tests}
}
\value{
Returns an estimated p-value for some variable.
}
\description{
This is a function to estimates the p-value
}
\author{
Nis Klausen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{nklau1414@student.sdu.dk} \cr
}
\keyword{chisq}

