% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task32.R
\name{Task32}
\alias{Task32}
\title{Task32}
\usage{
Task32(x, p = 1/length(x))
}
\arguments{
\item{x}{is the variables}

\item{p}{is the probability for the variables}
}
\value{
The test variable of a goodness of fit chi square test
}
\description{
This is a function to perform a chi square goodness of fit test
}
\author{
Nis Klausen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{nklau1414@student.sdu.dk} \cr
}
\keyword{chi}
\keyword{fit}
\keyword{goodnes}
\keyword{of}
\keyword{square}

