% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task41.R
\name{Task41}
\alias{Task41}
\title{Task41}
\usage{
Linear regression of some dependant and predictors
}
\arguments{
\item{A}{single criterion variable and some number of predictor variables in the formal criterion~predictors.}
}
\value{
Returns a list containing the same information as lm()
}
\description{
This is a funciton that takes some variables and performs linear regression on it. It use least squares to estimate the coeeficients.
}
\author{
Nis Klausen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{nklau1414@student.sdu.dk} \cr
}
\keyword{Linear}
\keyword{Regression}

