% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task51.R
\name{Task51}
\alias{Task51}
\title{Task51}
\usage{
Estimate density of a point.
}
\arguments{
\item{x}{is the set of variables.'}

\item{d}{is the point which the user wishes to estimate}

\item{h}{is the bandwidth}

\item{method}{is the method used.}
}
\value{
Returns the estimated density for either a point or the min, max, mean, median and the 1st and 3rd quantile if the user doesn't specify a point.
}
\description{
This is a function that takes some variables and returns an estimated density. If the user does not enter a point to be estimated the function returns the density for min, max, mean, median and the 1st and 3rd quantile. The user can choose the bandwidth and the method. The default method is the naive estimator defaulting Sturge's bandwidth. Alternative method is the Gaussian kernel which defaults Silverman's suggestion for Gaussian bandwidth.
}
\author{
Nis Klausen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{nklau1414@student.sdu.dk} \cr
}
\keyword{Density}
\keyword{Estimation}
\keyword{Gaussian}
\keyword{Naive}

