% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task52.R
\name{Task52}
\alias{Task52}
\title{PackageExam2016}
\usage{
Task52(x, n = 500, method = "naive", from = (min(x) + sd(x)/3),
  to = (max(x) - sd(x)/3), color = "skyblue")
}
\arguments{
\item{x}{is the variable}

\item{n}{is the number of points estimated. Defaults to 500}

\item{from}{is where the plot starts. Defaults to min(x) plus 1/3 of the standard deviation}

\item{to}{is where the plot ends. Defaults to max(x) minus 1/3 of the standard deviation}

\item{Method}{is the method of estimation, defaults to the naive estimator but also allows Gaussian kernel.}
}
\value{
Plot of some density estimation
}
\description{
This is a function that take a sample as input and returns a plot based on the function Task51
}
\author{
Nis Klausen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{nklau1414@student.sdu.dk} \cr
}
\keyword{Plot}
\keyword{density}
\keyword{estimation}

