% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task61.R
\name{Task61}
\alias{Task61}
\title{Task61}
\usage{
Task61(p, k, n = 10000)
}
\arguments{
\item{p}{is the transition probability matrix}

\item{k}{is the initial state.}

\item{n}{is the desired steps in the, defaulting to 10000 if nothing is entered.}
}
\value{
A list containing the path through the markov chain and the stationary probability of the various states
}
\description{
This is a function to go through a markov chain and finding the stationary probabilities.
}
\author{
Nis Klausen \cr
Department of mathematics and computer science (IMADA) \cr
University of southern Denmark, SDU, Odense \cr
\email{nklau1414@student.sdu.dk} \cr
}
\keyword{Carlo}
\keyword{Chain}
\keyword{Markov}
\keyword{Monte}
\keyword{Probability}
\keyword{Stationary}

